Revert "feat: support dynamic RPC URL override in keychain (#2052)"
This reverts commit e7d5562178fe037e47846bbcdd453b11947950b7.
